Refactor: Ajustes nas rotas e no reademe
