+++ src/controllers/registUser.js
@@ -1,8 +1,8 @@
 const bcrypt = require('bcrypt');
 
-const checkEmailRegistered = require('../utils.js/checkEmailRegistered');
-const validateUserDataRegister = require('../utils.js/validateUserDataRegister');
-const insertUserIntoDatabase = require('../utils.js/insertUserIntoDatabase');
+const checkEmailRegistered = require('../utils/checkEmailRegistered');
+const validateUserDataRegister = require('../utils/validateUserDataRegister');
+const insertUserIntoDatabase = require('../utils/insertUserIntoDatabase');
 
 const userRegistration = async (req, res) => {
     const { nome, email, senha } = req.body;
